Initial Trial of Recording White Noise
This is a project with a "Tune" and "Playback" option.  Essentially all i got was the phone to play white noise and then to get a recording of it. The playback button currently is supposed to call the playRecording() function, but it doesn't seem to be working. If we could just gain access/find where that file/recording is, we could then start on the actual tuning aspect of the project.